ajustes para melhoria de código
