+++ src/app/auth/components/login/login.component.ts
@@ -4,7 +4,7 @@ import { select, Store } from '@ngrx/store';
 import { Observable } from 'rxjs';
 import { PersistanceService } from 'src/app/shared/services/persistance/persistance.service';
 import { loginAction } from '../../store/actions/login.actions';
-import { selectIsSubmitting, selectValidationErrors } from '../../store/selectors/auth-selectors';
+import { isSubmittingSelector, validationErrorsSelector } from '../../store/selectors/auth-selectors';
 import { IBackendErrors } from '../../types/backend-errors-interface';
 import { IRegisterRequest } from '../../types/register-request.interface';
 
@@ -38,8 +38,8 @@ export class LoginComponent implements OnInit {
   }
 
   private initializeValues(): void{
-    this.isSubmitting$ = this.store.pipe(select(selectIsSubmitting));
-    this.backendErrors$ = this.store.pipe(select(selectValidationErrors));
+    this.isSubmitting$ = this.store.pipe(select(isSubmittingSelector));
+    this.backendErrors$ = this.store.pipe(select(validationErrorsSelector));
   }
 
   onSubmit(): void{
