+++ src/app/shared/modules/top-bar/components/top-bar/top-bar.component.html
@@ -5,7 +5,7 @@
       <li class="nav-item">
         <a [routerLink]="['/']" class="nav-link">Home</a>
       </li>
-      <ng-container *ngIf="(registerIsLoggedIn$ | async) || (loginIsLoggedIn$ | async) || (currentUser$ | async)">
+      <ng-container *ngIf="isLoggedIn$ | async">
         <li class="nav-item">
           <a [routerLink]="['/articles/new']" class="nav-link">
             <i class="ion-compose"></i>
@@ -27,7 +27,7 @@
           </a>
         </li>
       </ng-container>
-      <ng-container *ngIf="!(currentUser$ | async)">
+      <ng-container *ngIf="isAnonymous$ | async">
         <li class="nav-item">
           <a [routerLink]="['/login']" class="nav-link">
             Sign in
