+++ src/app/shared/modules/top-bar/components/top-bar/top-bar.component.ts
@@ -2,22 +2,12 @@ import { Component, OnInit } from '@angular/core';
 import { Store, select } from '@ngrx/store';
 import { Observable } from 'rxjs';
 import {
-  selectCurrentUser as registerUserSelector,
-  selectIsAnonymous as registerAnonymousSelector,
-  selectIsLoggedIn as registerIsLoggedInSelector
+  currentUserSelector,
+  isAnonymousSelector,
+  isLoggedInSelector
 } from 'src/app/auth/store/selectors/auth-selectors';
-import {
-  selectCurrentUser as loginUserSelector,
-  selectIsAnonymous as loginAnonymousSelector,
-  selectIsLoggedIn as loginIsLoggedInSelector
-} from 'src/app/auth/store/selectors/login-selectors';
-import {
-  selectCurrentUser as userSelector,
-  selectIsAnonymous as userAnonymousSelector,
-  selectIsLoggedIn as userIsLoggedInSelector
-} from 'src/app/auth/store/selectors/user-selectors';
+import { IAuthState } from 'src/app/auth/types/auth-state.interface';
 import { ICurrentUser } from 'src/app/shared/types/current-user.interface';
-//import { State as LoginState } from '../../../../../auth/store/reducers/login-reducer'
 
 @Component({
   selector: 'app-top-bar',
@@ -26,31 +16,18 @@ import { ICurrentUser } from 'src/app/shared/types/current-user.interface';
 })
 export class TopBarComponent implements OnInit {
 
-  registerIsLoggedIn$: Observable<boolean>;
-  registerIsAnonymous$: Observable<boolean>;
-  registerCurrentUser$: Observable<ICurrentUser | null>;
-
-  loginIsLoggedIn$: Observable<boolean>;
-  loginIsAnonymous$: Observable<boolean>;
-  loginCurrentUser$: Observable<ICurrentUser | null>;
-
+  isLoggedIn$: Observable<boolean>;
+  isAnonymous$: Observable<boolean>;
   currentUser$: Observable<ICurrentUser | null>;
 
-
   constructor(
-    private store: Store
+    private store: Store<IAuthState>
   ) { }
 
   ngOnInit(): void {
-    this.registerIsLoggedIn$ = this.store.pipe(select(registerIsLoggedInSelector));
-    this.registerIsAnonymous$ = this.store.pipe(select(registerAnonymousSelector));
-    this.registerCurrentUser$ = this.store.pipe(select(registerUserSelector));
-
-    this.loginIsLoggedIn$ = this.store.pipe(select(loginIsLoggedInSelector));
-    this.loginIsAnonymous$ = this.store.pipe(select(loginAnonymousSelector));
-    this.loginCurrentUser$ = this.store.pipe(select(loginUserSelector));
-
-    this.currentUser$ = this.store.pipe(select(userSelector))
+    this.isLoggedIn$ = this.store.pipe(select(isLoggedInSelector));
+    this.isAnonymous$ = this.store.pipe(select(isAnonymousSelector));
+    this.currentUser$ = this.store.pipe(select(currentUserSelector));
   }
 
 }
